requirements.txt
shapelets-solo==0.3.1